Moved libultra code that is inside "code" into its own folder, decompiled a bunch of files.
