+++ include/ultra64/rsp.h
@@ -3,6 +3,8 @@
 
 /* SP Registers */
 
+#define SP_MEM_ADDR_REG 0xA4040000
+#define SP_DRAM_ADDR_REG 0xA4040004
 #define SP_STATUS_REG 0xA4040010
 #define SP_PC_REG 0xA4080000
