+++ include/variables.h
@@ -202,7 +202,7 @@ extern int D_80009450;
 //extern ? D_8000AD00;
 //extern ? D_8000AD10;
 //extern ? D_8000AD38;
-//extern ? D_8000AD54;
+extern u32 D_8000AD54;
 //extern ? D_8000AD60;
 //extern ? D_8000AD7C;
 //extern ? D_8000AD80;
