+++ spec
@@ -387,11 +387,11 @@ beginseg
     include "build/asm/code_80103010.o"
     include "build/asm/sqrtf.o"
     include "build/asm/code_801031F0.o"
-    include "build/asm/code_80103B30.o"
+    include "build/src/libultra_code/code_80103B30.o"
     include "build/asm/code_80103B60.o"
     include "build/asm/code_80103BB0.o"
     include "build/asm/code_80103DC0.o"
-    include "build/asm/code_80104140.o"
+    include "build/src/libultra_code/code_80104140.o"
     include "build/asm/code_80104160.o"
     include "build/src/libultra/osDpGetStatus.o"
     include "build/src/libultra/osDpSetStatus.o"
@@ -413,14 +413,14 @@ beginseg
     include "build/asm/code_80105F40.o"
     include "build/asm/code_80106170.o"
     include "build/asm/code_801062E0.o"
-    include "build/asm/__osSpGetStatus.o"
-    include "build/asm/__osSpSetStatus.o"
+    include "build/src/libultra_code/__osSpGetStatus.o"
+    include "build/src/libultra_code/__osSpSetStatus.o"
     include "build/asm/osWritebackDCacheAll.o"
-    include "build/asm/code_801064C0.o"
+    include "build/src/libultra_code/code_801064C0.o"
     include "build/asm/code_801064E0.o"
     include "build/asm/code_80106550.o"
-    include "build/src/code_80106760.o"
-    include "build/src/code_801067A0.o"
+    include "build/src/libultra_code/code_80106760.o"
+    include "build/src/libultra_code/code_801067A0.o"
     include "build/asm/code_801067D0.o"
     include "build/asm/code_801067E0.o"
     include "build/asm/code_801067F0.o"
