+++ src/libultra_code/__osSpGetStatus.c
@@ -1,0 +1,7 @@
+#include <ultra64.h>
+#include <global.h>
+
+u32 __osSpGetStatus()
+{
+    return *(u32*)(SP_STATUS_REG);
+}
