+++ src/libultra_code/__osSpSetStatus.c
@@ -1,0 +1,7 @@
+#include <ultra64.h>
+#include <global.h>
+
+void __osSpSetStatus(u32 a0)
+{
+    *(u32*)(SP_STATUS_REG) = a0;
+}
