+++ src/libultra_code/code_80101B40.c.ignore
@@ -1,0 +1,22 @@
+#include <ultra64.h>
+#include <global.h>
+
+void func_80101B40(u32* a0)
+{
+    a0[0] = 0x3F800000; // Float?
+    a0[1] = 0;
+    a0[2] = 0;
+    a0[3] = 0;
+    a0[4] = 0;
+    a0[5] = 0;
+    a0[6] = 0x3F800000;
+    a0[7] = 0;
+    a0[8] = 0;
+    a0[9] = 0;
+    a0[10] = 0x3F800000;
+    a0[11] = 0;
+    a0[12] = 0;
+    a0[13] = 0;
+    a0[14] = 0;
+    a0[15] = 0x3F800000;
+}
