+++ src/libultra_code/code_80104140.c
@@ -1,0 +1,8 @@
+#include <ultra64.h>
+#include <ultra64/os_internal.h>
+#include <global.h>
+
+u32 func_80104140()
+{
+    return D_8000AD4C;
+}
