+++ src/libultra_code/code_801064C0.c
@@ -1,0 +1,7 @@
+#include <ultra64.h>
+#include <global.h>
+
+u32 func_801064C0()
+{
+    return D_8000AD54;
+}
