+++ src/z_debug_display.c.ignore
@@ -1,0 +1,9 @@
+#include <ultra64.h>
+#include <global.h>
+
+extern u32 D_8015FCB0;
+
+void func_80063E90()
+{
+    D_8015FCB0 = 0;
+}
